tests: remove and set annotations on VM creation
Currently, our tests create a VM and then use
"tests.RetryWithMetadataIfModified()" to update
the VM with annotations.

This process involves creating the VM object in
a non-running state, patching it with annotations,
and then starting it.

Instead, setting annotations during VM creation
simplifies the process and improves efficiency.

Additionally, remove irrelevant functional test
that was testing carrying annotations from VM
to VMI and ignoring "kubevirt.io" and
"kubernetes.io" labels. This filtering and
functionality was removed in PR #3110.

Instead, add a new unit test that tests the
scenario where VM's annotations can not be
carried to VMI annotations.

Signed-off-by: Ben Oukhanov <boukhanov@redhat.com>