fix: convert login time to seconds
Libvirt API returns milliseconds of the
loginTime, and previously KubeVirt
API was using QEMU directly that
represents seconds only.

This conversion from milliseconds to
seconds is essential to keep consistency
of what is expected from the KubeVirt
API in the response of loginTime.

Also update the description of the
loginTime in the KubeVirt API
to ensure that user understands
the returned value of it.

Signed-off-by: Ben Oukhanov <boukhanov@redhat.com>